Move all original content to an appendix
We don't want to lose any of the original content but it is not in
a terribly useful form as is.  So, move all of that content to an
appendix; as it is needed for reuse, or it is determined that is no
longer needed, it will be removed from the appendix, with the goal
being to ultimately being able to delete the appendix completely
since it will be empty.

Signed-off-by: Al Stone <ahs3@redhat.com>